feat: Allow URLs to accept both an artist's name and an artist's database id
Implement additional routes to either accept strings or ints to query for an artist in the database. Integers will correspond to the artist's database id, while a string will correspond to an artist's url_name column in the database. A helper function will parse the query and return a single database entity regardless of whether or not the query is by name or id.